+++ server/src/services/handleRecipeSearch.js
@@ -1,0 +1,16 @@
+import { User } from '../models/index.js'
+import RecipeSerializer from '../serializers/RecipeSerializer.js'
+
+const handleRecipeSearch = async (req) => {
+  const keyword = req.query.keyword.toLowerCase()
+  const user = await User.query().findById(req.user.id)
+  const recipes = await user
+    .$relatedQuery('recipes')
+    .where('name', 'ILike', `%${keyword}%`)
+    .orWhere('tags', '@>', [`${keyword}`])
+    .orWhere('notes', 'ILike', `%${keyword}%`)
+  const serializedRecipes = recipes.map(recipe => RecipeSerializer.getSummary(recipe))
+  return {serializedRecipes, keyword}
+}
+
+export default handleRecipeSearch
